requirements.txt
pytest
sqlAlchemy
pandas
psycopg2-binary
xlrd>=1.0.0